Set up a branch for the current development version of udpkg.
The problem with this branch of udpkg is that is has many changes, and
little testing. Some of the changes since the last release have already
been found to intrduce new bugs, and been reverted. It is thus not suitable
to be used for the sarge release, as version 0.025 of udpkg is basically
fine for that release.