tagging version 0.138
This commit contains a diff in the da.po file as a result of a cleanup of
corruption that started with revision 55901. The diff was included during
the cleanup to keep the tagged version identical to what was uploaded to
the archive at the time.
See also: http://lists.debian.org/debian-boot/2009/06/msg00081.html