Work around bug in rdate when system date is smaller than the epoch
Merged from trunk.

If the current system date is before the epoch, rdate fails to set it to
the date obtained using NTP (#502336). Work around this by adding a small
C program (set-date-epoch) that sets the date to the epoch (1-1-1970) before
calling rdate, but only does so if the current date is smaller.