WIP: Rewrite prepare-release-announce
This is a proof of concept, switching from parsing build logs to diffing
Sources files between the archive serials recorded by debian-cd during
the build of installation images. This should cover more ground (esp.
sources shipping udebs that are only used at run-time, not seen in build
logs).

This relies on snapshot.debian.org's having a timestamp for each archive
serial.

Add the moment: The lookup needs to be done manually, and timestamps for
D-I Buster Alpha 3 are hardcoded.